+++ Geometry.cpp
@@ -6,18 +6,18 @@ Light::Light() {
   // Конструктор класса Light
 }
 
-Sphere::Sphere(Vector3d center, double radius, wxColour color, double specular,
-               double reflective)
+Sphere::Sphere(Vector3d center, float radius, wxColour color, float specular,
+               float reflective)
     : center(center),
       radius(radius),
       color(color),
       specular(specular),
       reflective(reflective) {}
 
-AmbientLight::AmbientLight(double intensity) : intensity(intensity) {}
+AmbientLight::AmbientLight(float intensity) : intensity(intensity) {}
 
-PointLight::PointLight(double intensity, Vector3d position)
+PointLight::PointLight(float intensity, Vector3d position)
     : intensity(intensity), position(position) {}
 
-DirectionalLight::DirectionalLight(double intensity, Vector3d directional)
+DirectionalLight::DirectionalLight(float intensity, Vector3d directional)
     : intensity(intensity), directional(directional) {}
